@@ -1,16 +1,14 @@
-import {async, TestBed} from '@angular/core/testing';
+import {async, ComponentFixture, TestBed} from '@angular/core/testing';
 import {RouterTestingModule} from '@angular/router/testing';
 import {FormsModule} from '@angular/forms';
 import {Observable} from 'rxjs/Observable';
 import {AppComponent} from './app.component';
 import {HeaderComponent} from './layout-components/header/header.component';
 import {FooterComponent} from './layout-components/footer/footer.component';
-import {ComponentFixture} from '@angular/core/testing';
 import {NavigationBarComponent} from './layout-components/navigation-bar/navigation-bar.component';
 import {WordpressPagesService} from './shared/wordpress-pages-service/wordpress-pages.service';
 import {GoogleAnalyticsService} from './shared/analytics/google-analytics.service';
-import {SVGCacheService} from 'ng-inline-svg';
-import {InlineSVGModule} from 'ng-inline-svg';
+import {InlineSVGModule, SVGCacheService} from 'ng-inline-svg';
 import {HttpClientTestingModule} from '@angular/common/http/testing';
 import {OneTimePopupComponent} from './layout-components/one-time-popup/one-time-popup.component';
 import {PopupComponent} from './shared/popup/popup.component';
@@ -19,6 +17,8 @@ import {WordpressMeasuresService} from './shared/wordpress-measures-service/word
 import {NavBarSuboptionComponent} from "./layout-components/navigation-bar/nav-bar-suboption/nav-bar-suboption.component";
 import {UserStateService} from "./shared/user-state-service/user-state-service";
 import {NeedHelpComponent} from "./shared/need-help/need-help.component";
+import {SearchBarComponent} from "./layout-components/search-bar/search-bar.component";
+import {RecommendationsService} from "./shared/recommendations-service/recommendations.service";
 
 describe('AppComponent', () => {
     let fixture: ComponentFixture<AppComponent>;
@@ -43,6 +43,7 @@ describe('AppComponent', () => {
                 PopupComponent,
                 NavBarSuboptionComponent,
                 NeedHelpComponent,
+                SearchBarComponent
             ],
             imports: [RouterTestingModule, FormsModule, InlineSVGModule, HttpClientTestingModule],
             providers: [
@@ -52,6 +53,7 @@ describe('AppComponent', () => {
                 {provide: SVGCacheService, useValue: {setBaseUrl: () => {}}},
                 {provide: CookieService, useValue: mockCookieService},
                 {provide: UserStateService, useValue: mockUserStateService},
+                {provide: RecommendationsService, useValue: {}},
             ]
         }).compileComponents();
     }));
