@@ -3,7 +3,6 @@ import {RouterTestingModule} from '@angular/router/testing';
 import {CommonModule} from '@angular/common';
 import {FormsModule} from '@angular/forms';
 import {Observable} from 'rxjs/Observable';
-import {InlineSVGModule} from 'ng-inline-svg';
 
 import {HomeImprovementsComponent} from './home-improvements.component';
 import {LargeVideoCardComponent} from '../../shared/large-video-card/large-video-card.component';
@@ -18,13 +17,16 @@ import {QuestionContentService} from '../../shared/question-content/question-con
 import {PostcodeDetails} from '../../shared/postcode-epc-service/model/postcode-details';
 import {PostcodeEpcService} from '../../shared/postcode-epc-service/postcode-epc.service';
 import {WordpressPagesService} from '../../shared/wordpress-pages-service/wordpress-pages.service';
+import {InlineSVGModule} from 'ng-inline-svg';
 import {StaticMeasureCardComponent} from '../static-measure-card/static-measure-card.component';
 import {DataCardComponent} from '../../shared/data-card/data-card.component';
 import {PostcodeApiService} from "../../shared/postcode-epc-service/postcode-api-service/postcode-api.service";
 import {PostcodeBasicDetailsResponse} from '../../shared/postcode-epc-service/model/response/postcode-basic-details-response';
 import { PopupComponent } from '../../shared/popup/popup.component';
 import {LatestNewsSectionComponent} from '../../shared/latest-news-section/latest-news-section.component';
+import {SearchBarComponent} from "../../layout-components/search-bar/search-bar.component";
 import {NavBarSuboptionComponent} from "../../layout-components/navigation-bar/nav-bar-suboption/nav-bar-suboption.component";
+import {GoogleAnalyticsService} from "../../shared/analytics/google-analytics.service";
 
 describe('HomeImprovementsComponent', () => {
     let component: HomeImprovementsComponent;
@@ -60,13 +62,14 @@ describe('HomeImprovementsComponent', () => {
                 StaticMeasureCardComponent,
                 DataCardComponent,
                 PopupComponent,
+                SearchBarComponent,
                 NavBarSuboptionComponent,
             ],
             imports: [
                 CommonModule,
                 FormsModule,
                 RouterTestingModule,
-                InlineSVGModule,
+                InlineSVGModule
             ],
             providers: [
                 ResponseData,
@@ -74,6 +77,7 @@ describe('HomeImprovementsComponent', () => {
                 {provide: PostcodeEpcService, useValue: postcodeEpcServiceStub},
                 {provide: WordpressPagesService, useValue: {getLatestPages: () => Observable.of([])}},
                 {provide: PostcodeApiService, useValue: postcodeApiServiceStub},
+                GoogleAnalyticsService,
             ]
         })
             .compileComponents();
