@@ -1,4 +1,4 @@
-import {Component, Input} from '@angular/core';
+import {Component, Input, Output, EventEmitter} from '@angular/core';
 import {WordpressPage} from '../wordpress-pages-service/wordpress-page';
 
 @Component({
@@ -8,4 +8,9 @@ import {WordpressPage} from '../wordpress-pages-service/wordpress-page';
 })
 export class LatestNewsCardComponent {
     @Input() page: WordpressPage;
+    @Output() onArticleClicked: EventEmitter<null> = new EventEmitter<null>();
+
+    onClick() {
+        this.onArticleClicked.emit();
+    }
 }
