@@ -1,6 +1,7 @@
-import { Component, OnInit } from '@angular/core';
+import {Component, OnInit, Input} from '@angular/core';
 import {WordpressPage} from '../wordpress-pages-service/wordpress-page';
 import {WordpressPagesService} from '../wordpress-pages-service/wordpress-pages.service';
+import {GoogleAnalyticsService} from "../analytics/google-analytics.service";
 
 @Component({
     selector: 'app-latest-news-section',
@@ -10,11 +11,21 @@ import {WordpressPagesService} from '../wordpress-pages-service/wordpress-pages.
 export class LatestNewsSectionComponent implements OnInit {
     latestNews: WordpressPage[];
 
-    constructor(private pageService: WordpressPagesService) {
+    @Input() googleAnalyticsCategory: string;
+
+    constructor(private pageService: WordpressPagesService,
+                private googleAnalyticsService: GoogleAnalyticsService) {
     }
 
     ngOnInit() {
         this.pageService.getLatestPages()
             .subscribe(pages => this.latestNews = pages);
+    }
+
+    sendEventToAnalytics(articleIndex: number) {
+        if (this.googleAnalyticsCategory) {
+            const eventName = `article${articleIndex + 1}_clicked`;
+            this.googleAnalyticsService.sendEvent(eventName, this.googleAnalyticsCategory);
+        }
     }
 }
