@@ -1,9 +1,11 @@
 import {async, ComponentFixture, TestBed} from '@angular/core/testing';
+import {RouterTestingModule} from '@angular/router/testing';
 import {InlineSVGModule} from 'ng-inline-svg';
 
 import {NeedHelpComponent} from './need-help.component';
 import {By} from '@angular/platform-browser';
 import {UserStateService} from "../user-state-service/user-state-service";
+import {GoogleAnalyticsService} from "../analytics/google-analytics.service";
 
 describe('NeedHelpComponent', () => {
     let component: NeedHelpComponent;
@@ -16,8 +18,14 @@ describe('NeedHelpComponent', () => {
     beforeEach(async(() => {
         TestBed.configureTestingModule({
             declarations: [ NeedHelpComponent ],
-            imports: [InlineSVGModule],
-            providers: [{provide: UserStateService, useValue: userStateServiceStub}]
+            imports: [
+                RouterTestingModule,
+                InlineSVGModule,
+            ],
+            providers: [
+                {provide: UserStateService, useValue: userStateServiceStub},
+                GoogleAnalyticsService,
+            ]
         })
             .compileComponents();
     }));
