@@ -1,5 +1,6 @@
 import {Component, Input, OnInit} from '@angular/core';
 import {UserStateService} from "../user-state-service/user-state-service";
+import {GoogleAnalyticsService} from "../analytics/google-analytics.service";
 
 @Component({
     selector: 'app-need-help',
@@ -11,10 +12,23 @@ export class NeedHelpComponent implements OnInit {
     expanded: boolean;
     reference: string = "";
 
-    constructor(private userStateService: UserStateService) {
+    constructor(private userStateService: UserStateService,
+                private googleAnalyticsService: GoogleAnalyticsService) {
     }
 
     async ngOnInit() {
         this.reference = await this.userStateService.getSessionReference();
     }
+
+    toggleExpanded() {
+        this.expanded = !this.expanded;
+        if (this.expanded) {
+            this.sendEventToAnalytics('help-button_opened');
+        }
+    }
+
+    sendEventToAnalytics(eventName: string) {
+        this.googleAnalyticsService.sendEvent(eventName, 'assisted-digital');
+    }
+
 }
