@@ -1,5 +1,5 @@
-import * as parse from 'url-parse';
 import {RecommendationStepResponse} from '../energy-saving-measure-content-service/recommendation-step-response';
+import {WordpressPagesService} from "../wordpress-pages-service/wordpress-pages.service";
 
 export class RecommendationStep {
     headline: string;
@@ -21,7 +21,9 @@ export class RecommendationStep {
                 return {
                     buttonText: link.button_text,
                     isExternalLink: link.is_external_link,
-                    route: parse(link.linked_page).pathname,
+                    // The linked_page is a Wordpress Post URL, like:
+                    //   http://admin-site/category/slug
+                    route: WordpressPagesService.getRouteForWordpressPageFromPageUrl(link.linked_page),
                     externalLink: link.external_link
                 };
             });
