@@ -1,5 +1,6 @@
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
+import {InlineSVGModule} from 'ng-inline-svg';
 
 import {SharedModule} from "../shared/shared.module";
 import {RoutingModule} from "../app-routing.module";
@@ -15,7 +16,8 @@ import {YourHomeComponent} from './your-home.component';
     imports: [
         SharedModule,
         CommonModule,
-        RoutingModule
+        RoutingModule,
+        InlineSVGModule
   ]
 })
 export class YourHomeModule {
